401 and float fix

+++ internal/storage/db.go
@@ -17,9 +17,9 @@ type DB struct {
 }
 
 type Order struct {
-	Number     string `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual"`
+	Number     string  `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual"`
 	UserID     string
 	UploadedAt time.Time `json:"uploaded_ats"`
 }
